Nav.css, Nav.js: fix width of nav bar
otherwise, the width is 100% for the content. This means that if
I have a window of width 500, the content will have a width of 500,
and with both sides having a 20px padding it will have an actual
width of 540, and the right hand side will go outside the page.

The fix is to add a fixed wrapper with no border, padding or margin,
and we will give it a width of 100%.
Inside this wrapper, the original nav can exist with an automatic
width. This allows the inner element to automatically compute its
width.